drm/i915/gvt: Expand direct display port assignment mask to 64bit
i915 driver defines a larger I915_MAX_PORTS which can't be hold in 32bit
integer as the direct display port availability, selected mask and
assignment mask. In GEN9, the port number couldn't be so large, and
current implementation requries a 8-bit mask for each vGPU to hold the
port assignment.

Expand the availability and ownership to 64bit so that host port status
can be displayed correctly. Still keep using 8-bit mask for port
assignment.

Signed-off-by: Colin Xu <colin.xu@intel.com>